image_types_qcom: copy spinor firmware if available
Spinor firmware is needed on iq-x7181 based devices, and recently
included as part of hamoa sprint one release, deploy them to generate
the flat build for spinor flash. Meanwhile, copy the programer outside
of spinor folder to be part of NVME flat build.

+++ classes-recipe/image_types_qcom.bbclass
@@ -86,7 +86,7 @@ create_qcomflash_pkg() {
         install -m 0644 ${rawpg} .
     done
 
-    if [ -n "${QCOM_CDT_FILE}" ]; then
+    if [ -n "${QCOM_CDT_FILE}" && -e "${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/${QCOM_CDT_FILE}.bin" ]; then
         install -m 0644 ${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/${QCOM_CDT_FILE}.bin cdt.bin
     fi
 
@@ -114,6 +114,13 @@ create_qcomflash_pkg() {
     if [ -d "${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/sail_nor" ]; then
         install -d sail_nor
         find "${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/sail_nor" -maxdepth 1 -type f -exec install -m 0644 {} sail_nor \;
+    fi
+
+    # spinor firmware
+    if [ -d "${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/spinor" ]; then
+        install -d spinor
+        find "${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/spinor" -maxdepth 1 -type f -exec install -m 0644 {} spinor \;
+        find "${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/spinor" -maxdepth 1 -type f -name 'xbl_s_devprg_ns.melf' -exec install -m 0644 {} . \;
     fi
 
     # Create symlink to ${QCOMFLASH_DIR} dir
